+++ mopidy_local_images/library.py
@@ -12,22 +12,11 @@ import struct
 from mopidy import local
 from mopidy.audio import scan
 from mopidy.exceptions import ExtensionError
+from mopidy.local import translator
 
 import uritools
 
 from . import Extension
-
-# changed in Mopidy v1.1
-try:
-    from mopidy.local.translator import local_uri_to_file_uri
-except ImportError:
-    from mopidy.local.translator import local_track_uri_to_file_uri \
-        as local_uri_to_file_uri
-try:
-    from mopidy.local.translator import local_uri_to_path
-except ImportError:
-    from mopidy.local.translator import local_track_uri_to_path \
-        as local_uri_to_path
 
 logger = logging.getLogger(__name__)
 
@@ -119,7 +108,7 @@ class ImageLibrary(local.Library):
 
     def add(self, track, tags=None, duration=None):
         if track.album and track.album.name:  # FIXME: album required
-            uri = local_uri_to_file_uri(track.uri, self.media_dir)
+            uri = translator.local_uri_to_file_uri(track.uri, self.media_dir)
             try:
                 if tags is None:
                     images = self._extract_images(track.uri, self._scan(uri))
@@ -188,7 +177,9 @@ class ImageLibrary(local.Library):
                 images.add(self._get_or_create_image_file(None, data))
             except Exception as e:
                 logger.warn('Error extracting images for %r: %r', uri, e)
-        dirname = os.path.dirname(local_uri_to_path(uri, self.media_dir))
+        # look for external album art
+        path = translator.local_uri_to_path(uri, self.media_dir)
+        dirname = os.path.dirname(path)
         for pattern in self.patterns:
             for path in glob.glob(os.path.join(dirname, pattern)):
                 try: